Work with correct usage, cannot chain yet
calculator works if used properly; '=' works after entering
two numbers and an operator. does not work yet if an operator is
entered instead of '='.
For example, 2 + 8 = 10 will work, but 2 + 7 + 1 will have incorrect result.
'.' can be used to make decimal numbers as well.
Delete and AC buttons do not work yet.